PIN-206: moved Setor DTO into DTO namespace
